- New trimesh box collision function from scratch. - Opcode internal support for double precision vertices in meshinterface. - Opcode vc++ project file modified. - Drawstuff now has double precision dsDrawTriangle() - Fixed some small trimesh problems.

+++ include/drawstuff/drawstuff.h
@@ -141,7 +141,7 @@ void dsDrawBoxD (const double pos[3], const double R[12],
 void dsDrawSphereD (const double pos[3], const double R[12],
 		    const float radius);
 void dsDrawTriangleD (const double pos[3], const double R[12],
-		      const float *v0, const float *v1, const float *v2, int solid);
+		      const double *v0, const double *v1, const double *v2, int solid);
 void dsDrawCylinderD (const double pos[3], const double R[12],
 		      float length, float radius);
 void dsDrawCappedCylinderD (const double pos[3], const double R[12],
